requirements-dev.txt
aiofiles==0.4.0
asyncio==3.4.3
black==19.10b0
httpx==0.11.1
Jinja2==2.11.1
parse==1.14.0
pre-commit==2.2.0
python-multipart==0.0.5
pytest==5.3.5
pytest-asyncio==0.10.0
pytest-cov==2.8.1
PyYAML==5.3.1
uvicorn==0.11.3
websocket-client==0.57.0

# documentation
mkdocs==1.1
mkdocs-material==4.6.3
